refactor(graph): move read only variable pool and runtime state to runtime package
