+++ api/core/workflow/runtime/__init__.py
@@ -1,11 +1,14 @@
 from .graph_runtime_state import GraphRuntimeState
 from .graph_runtime_state_protocol import ReadOnlyGraphRuntimeState, ReadOnlyVariablePool
+from .read_only_wrappers import ReadOnlyGraphRuntimeStateWrapper, ReadOnlyVariablePoolWrapper
 from .variable_pool import VariablePool, VariableValue
 
 __all__ = [
     "GraphRuntimeState",
     "ReadOnlyGraphRuntimeState",
+    "ReadOnlyGraphRuntimeStateWrapper",
     "ReadOnlyVariablePool",
+    "ReadOnlyVariablePoolWrapper",
     "VariablePool",
     "VariableValue",
 ]
